NEW: CPWindow -layoutIfNeeded, -setNeedsLayout
setNeedsLayout adds CPWindow to the stack of layoutObjects in
_CPDisplayServer. _CPDisplayServer runLoop calls -layoutIfNeeded ->
-layout and runs the updateConstraintsIfNeeded pass and solves the
frames for the full hierarchy.
Currently setNeedsLayout is explicitly called when we tell the window
to engage auto layout (in -setAutolayoutEnabled:). In the auto layout
examples, this is done in awakeFromCib, i.e. after the view hierarchy
is setup. TODO: call setNeedsLayout in the FMW when all views are added.