Renames and fixes
FIXED: CPView -setFrameSize: is called when we update the view frame
with the solver results. As opposed to autosize, we don’t want to send
the receiver to the display server (setNeedsLayout) in this method
because the subviews are already resized by the solver.

FIXED: When a view has a dirty constraint or a dirty frame, it informs
the superview recursively. When updateConstraintForSubtree or
layoutSubtreeIfNeeded walk down the tree, the flag is just here to
select only the dirty branches in the hierarchy.
After this commit, when the superview is already marked as the ancestor
of a dirty view, we break because we assume the parents where already
marked by a previous call of the same method.

_CPViewUpdateEngineFrame : expand args for perf.