CPLayoutRect reference item
Changed rules again: now the layout rect is a full citizen layout item.
A named layout rect generates anchors that have the layout rect as a
reference item.
The layout rect superitem (equivalent of superview) is specified by the
inItem: argument.
This means that the layout rect receives all engine notifications for
its anchors changes. Currently it does nothing upon this (the layout
rect is not part of the descending layout pass).